Accelerometer calibration in EEPROM for X4
If there are no calibration parameters in EEPROM, everything stays the same as before: calibration of
accelerometer and gyro is performed after binding with TX and data not saved in EEPROM.
4 slow LED blinks have been added to indicate that no EEPROM data has been found and a level surface is needed.

If parameters have been loaded from EEPROM, only gyro calibration is executed after
binding with the TX. No need to worry about a level surface when flying outside anymore.

To calibrate accelerometer and store the calibration data in EEPROM:
 * put quadcopter on level surface
 * quadcopter must be not armed
 * throttle stick in lowest position
 * move roll stick back and forth 3 times
Rotating LED pattern indicates the calibration process.
Data is stored in EEPROM (=data flash in Mini54) automatically.

When uploading a firmware containing new PID parameters make sure to erase the data flash, otherwise the firmware will still use the old values.
The PID paramaters are stored together with the accelerometer calibration.

Trying to keep the code for the other models working:
On all other  models (not Hubsan X4) EEPROM storage is still disabled.
It should be possible to execute the calibration using the roll stick.
No change for LED.
This is not tested.

Changes to be committed:
	modified:   src/bradwii.c
	modified:   src/config_X4.h
	modified:   src/eeprom.c
	modified:   src/imu.c
	modified:   src/imu.h
	modified:   src/serial.c