New flight mode "yaw hold"
In "yaw hold" mode the yaw PID controller works on the accumulated yaw errors
instead of just the last yaw error. So it controls the angle instead of the
yaw rate. If the quadcopter gets rotated away from its desired heading, it will not actively
try to rotate back. It is similar to compass mode, but without using a compass.
Yaw hold mode can be activated using the checkboxes. It is currently set to AUX2 low.
I don't know if this mode works in combination with acro modes.

PID control parameters for yaw have been changed again.

Changes to be committed:
	modified:   src/checkboxes.c
	modified:   src/checkboxes.h
	modified:   src/config_X4.c
	modified:   src/pilotcontrol.c